FLIN-78 Move files into correct folders
