+++ flinkfood-demo/src/main/java/org/flinkfood/schemas/customer/Customer.java
@@ -1,4 +1,4 @@
-package org.flinkfood.flinkjobs;
+package org.flinkfood.schemas.customer;
 import java.util.Date;
 
 public class Customer {
