+++ flinkfood-demo/src/main/java/org/flinkfood/schemas/customer/Customer_address.java
@@ -1,4 +1,4 @@
-package org.flinkfood.flinkjobs;
+package org.flinkfood.schemas.customer;
 
 public class Customer_address {
