+++ flinkfood-demo/src/main/java/org/flinkfood/schemas/customer/KafkaAddressSchema.java
@@ -1,12 +1,11 @@
-package org.flinkfood.flinkjobs;
+package org.flinkfood.schemas.customer;
+
+import java.io.IOException;
 
 import org.apache.flink.api.common.serialization.AbstractDeserializationSchema;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.json.JsonMapper;
-import java.io.IOException;
-import java.time.Instant;
 
-public class KafkaCustomerSchema extends AbstractDeserializationSchema<Customer> {
+public class KafkaAddressSchema extends AbstractDeserializationSchema<Customer_address> {
 
     private static final long serialVersionUID = 1L;
     private transient ObjectMapper objectMapper;
@@ -17,8 +16,8 @@ public class KafkaCustomerSchema extends AbstractDeserializationSchema<Customer>
     }
 
     @Override
-    public Customer deserialize(byte[] message) throws IOException {
-        return objectMapper.readValue(message, Customer.class);
+    public Customer_address deserialize(byte[] message) throws IOException {
+        return objectMapper.readValue(message, Customer_address.class);
     }
 
 }
