+++ flinkfood-demo/src/main/java/org/flinkfood/schemas/order/Order.java
@@ -1,4 +1,4 @@
-package org.flinkfood;
+package org.flinkfood.schemas.order;
 
 import java.sql.Date;
